+++ src/main/java/me/desht/pneumaticcraft/common/ai/DroneEntityAIPickupItems.java
@@ -1,7 +1,8 @@
 package me.desht.pneumaticcraft.common.ai;
 
 import me.desht.pneumaticcraft.api.drone.IDrone;
-import me.desht.pneumaticcraft.common.progwidgets.ProgWidgetPickupItem;
+import me.desht.pneumaticcraft.common.progwidgets.IItemPickupWidget;
+import me.desht.pneumaticcraft.common.progwidgets.ProgWidgetAreaItemBase;
 import me.desht.pneumaticcraft.common.util.IOHelper;
 import me.desht.pneumaticcraft.lib.Names;
 import net.minecraft.entity.Entity;
@@ -16,15 +17,21 @@ import java.util.List;
 
 public class DroneEntityAIPickupItems extends Goal {
     private final IDroneBase drone;
-    private final ProgWidgetPickupItem itemPickupWidget;
+    private final ProgWidgetAreaItemBase itemPickupWidget;
     private ItemEntity curPickingUpEntity;
     private final DistanceEntitySorter theNearestAttackableTargetSorter;
+    private final boolean canSteal;
 
-    public DroneEntityAIPickupItems(IDroneBase drone, ProgWidgetPickupItem progWidgetPickupItem) {
+    public DroneEntityAIPickupItems(IDroneBase drone, ProgWidgetAreaItemBase progWidgetPickupItem) {
         this.drone = drone;
         setMutexFlags(EnumSet.allOf(Flag.class)); // so it won't run along with other AI tasks.
         itemPickupWidget = progWidgetPickupItem;
         theNearestAttackableTargetSorter = new DistanceEntitySorter(drone);
+        if (progWidgetPickupItem instanceof IItemPickupWidget) {
+            canSteal = ((IItemPickupWidget) itemPickupWidget).canSteal();
+        } else {
+            throw new IllegalArgumentException("expecting a IItemPickupWidget!");
+        }
     }
 
     /**
@@ -40,7 +47,7 @@ public class DroneEntityAIPickupItems extends Goal {
         }
         pickableItems.sort(theNearestAttackableTargetSorter);
         for (Entity ent : pickableItems) {
-            if (ent.getPersistentData().getBoolean(Names.PREVENT_REMOTE_MOVEMENT) && !itemPickupWidget.canSteal()) {
+            if (ent.getPersistentData().getBoolean(Names.PREVENT_REMOTE_MOVEMENT) && !canSteal) {
                 continue;
             }
             ItemStack stack = ((ItemEntity) ent).getItem();